+++ synth-brute/src/impl_material.rs
@@ -1,7 +1,7 @@
 use crate::*;
 
 impl Material {
-    pub fn new(color: Color, quality: usize, shape: Shape) -> Self {
+    pub fn new(color: Color, quality: u32, shape: Shape) -> Self {
         Self {
             color,
             effect_value: quality,
